Avoid submitting data from disabled fields
This involves pulling out the disabled data, submitting the form, then
putting the disabled data back in. We need to do the dance to avoid
regressing the form to hard-wired defaults, which can and does differ
from bound defaults for disabled fields (e.g., URL prefix in Admin >
Settings, which comes from config).